Add GUI window ready for UI
