+++ Cargo.toml
@@ -7,5 +7,5 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-nannou = { git = "https://github.com/nannou-org/nannou", branch = "master" }
+nannou = "0.11"
 sacn = "0.4"
